Separate default config and config class + use config class in restream.py
